+++ v2/cgo.go
@@ -3,8 +3,6 @@ package webui
 /*
 #cgo CFLAGS: -Iwebui/include/
 #cgo CFLAGS: -DNDEBUG -DNO_CACHING -DNO_CGI -DUSE_WEBSOCKET -DCGO
-// TODO: TLS condition
-#cgo CFLAGS: -DNO_SSL
 #cgo darwin CFLAGS: -x objective-c
 
 #cgo darwin LDFLAGS: -framework WebKit -framework Cocoa
