+++ v2/cgo_no_tls.go
@@ -1,0 +1,6 @@
+//go:build !webui_tls
+
+package webui
+
+// #cgo CFLAGS: -DNO_SSL
+import "C"
